Add authentication and relationships

+++ app/Controllers/Http/TodoController.js
@@ -20,11 +20,16 @@ class TodoController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async index ({ request, response, view }) {
-    const todos = await Todo.all()
+  async index ({ request, auth, response, view }) {
+    // const todos = await Todo.all()
+    const todos = await Todo
+      .query()
+      .where('user_id', auth.user.id)
+      .fetch()
 
     return view.render('index', {
-      todos: todos.toJSON()
+      todos: todos.toJSON(),
+      name: auth.user.username
     })
   }
 
@@ -48,7 +53,7 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store ({ request, session, response }) {
+  async store ({ request, auth, session, response }) {
     // const rules = {
     //   addTodo: 'required|min:3',
     // }
@@ -69,6 +74,7 @@ class TodoController {
     // }
 
     const todo = await Todo.create({
+      user_id: auth.user.id,
       title: request.input('addTodo')
     })
 
@@ -97,8 +103,12 @@ class TodoController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async edit ({ params, request, response, view }) {
+  async edit ({ params, auth, request, response, view }) {
     const todo = await Todo.findOrFail(params.id)
+
+    if (auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
 
     return view.render('edit', { todo })
   }
@@ -111,8 +121,13 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async update ({ params, session, request, response }) {
+  async update ({ params, auth, session, request, response }) {
     const todo = await Todo.findOrFail(params.id)
+
+    if (auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
+
     todo.title = request.input('editTodo')
     todo.completed = request.input('completedCheck') === 'on' ? true : false
     await todo.save()
@@ -129,8 +144,13 @@ class TodoController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async destroy ({ params, session, request, response }) {
+  async destroy ({ params, auth, session, request, response }) {
     const todo = await Todo.findOrFail(params.id)
+
+    if (auth.user.id !== todo.user_id) {
+      return 'You do not have permission to do this'
+    }
+
     await todo.delete()
 
     session.flash({ successMessage: 'Todo was deleted successfully!' })
